+++ lib/index.ts
@@ -1,5 +1,5 @@
-export * from "./actions";
-export * from "./matchers";
-export * from "./selectors";
+export * from "./state/actions";
+export * from "./state/matchers";
+export * from "./state/selectors";
 
 export * from "./types";
