+++ lib/state/actions/updateTransactionLastChecked.ts
@@ -1,4 +1,4 @@
-import type { TransactionsState } from "../types";
+import type { TransactionsState } from "../../types";
 
 export interface UpdateTransactionLastCheckedPayload {
   chainId: number;
