+++ lib/state/matchers/__tests__/matchIsTransactionConfirmed.test.ts
@@ -4,33 +4,14 @@ import {
   MOCK_CHAIN_ID_1,
   MOCK_CONFIRMED_TRANSACTION,
   MOCK_TRANSACTION,
-} from "../../test-utils";
+} from "../../../test-utils";
 
-import type { TransactionsState } from "../../types";
+import type { TransactionsState } from "../../../types";
 
-import { matchIsTransactionPending } from "../matchIsTransactionPending";
+import { matchIsTransactionConfirmed } from "../matchIsTransactionConfirmed";
 
-describe("matchIsTransactionPending", () => {
-  it("should return true if the transaction does not have a receipt or a confirmed time", () => {
-    const chainId = MOCK_CHAIN_ID_1;
-
-    const transaction = MOCK_TRANSACTION;
-
-    const transactionsState: TransactionsState = {
-      [chainId]: {
-        [transaction.hash]: transaction,
-      },
-    };
-
-    const result = matchIsTransactionPending(
-      transactionsState,
-      chainId,
-      transaction.hash
-    );
-
-    expect(result).toBe(true);
-  });
-  it("should return false if the transaction has a receipt and a confirmed time", () => {
+describe("matchIsTransactionConfirmed", () => {
+  it("should return true if the transaction has a receipt and a confirmed time", () => {
     const chainId = MOCK_CHAIN_ID_1;
 
     const transaction = MOCK_CONFIRMED_TRANSACTION;
@@ -41,7 +22,26 @@ describe("matchIsTransactionPending", () => {
       },
     };
 
-    const result = matchIsTransactionPending(
+    const result = matchIsTransactionConfirmed(
+      transactionsState,
+      chainId,
+      transaction.hash
+    );
+
+    expect(result).toBe(true);
+  });
+  it("should return false if the transaction does not have a receipt or a confirmed time", () => {
+    const chainId = MOCK_CHAIN_ID_1;
+
+    const transaction = MOCK_TRANSACTION;
+
+    const transactionsState: TransactionsState = {
+      [chainId]: {
+        [transaction.hash]: transaction,
+      },
+    };
+
+    const result = matchIsTransactionConfirmed(
       transactionsState,
       chainId,
       transaction.hash
@@ -56,7 +56,7 @@ describe("matchIsTransactionPending", () => {
 
     const transactionsState: TransactionsState = {};
 
-    const result = matchIsTransactionPending(
+    const result = matchIsTransactionConfirmed(
       transactionsState,
       chainId,
       transaction.hash
