+++ lib/state/selectors/getAllChainTransactions.ts
@@ -1,6 +1,6 @@
-import type { ChainTransactionsState, TransactionsState } from "../types";
+import type { ChainTransactionsState, TransactionsState } from "../../types";
 
-import { getValueOrDefault } from "../utils/getValueOrDefault";
+import { getValueOrDefault } from "../../utils/getValueOrDefault";
 
 export function getAllChainTransactions(
   transactionsState: TransactionsState,
